+++ src/main/java/com/example/ContaGest/repository/InvoiceRepository.java
@@ -9,14 +9,14 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface InvoiceRepository extends JpaRepository<InvoiceModel,Long>{
+public interface InvoiceRepository extends JpaRepository<InvoiceModel,Integer>{
 
     @Query("SELECT f FROM InvoiceModel f WHERE f.client.userCI =:clientCI AND f.month =:month")
     List<InvoiceModel> findByClientIdAndMonth(@Param("clientCI") String clientCI,@Param("month") int month);
 
     @Query("SELECT f.id FROM InvoiceModel f WHERE f.client.userCI =:clientCI")
-    List<Long> findIdByClientCI (@Param("clientCI") String clientCI);
+    List<Integer> findIdByClientCI (@Param("clientCI") String clientCI);
 
     @Query("SELECT f.id FROM InvoiceModel f WHERE f.client.userCI =:clientCI AND f.month =:month")
-    List<Long> findIdByClientCiAndMonth (@Param("clientCI") String clientCI,@Param("month") int month);
+    List<Integer> findIdByClientCiAndMonth (@Param("clientCI") String clientCI,@Param("month") int month);
 }
